Improve checking for presence of submodule
As libdivecomputer and subsurface are submodules, they would always
be present as folder on git cloning. So, on new install, this would
always be false and the submodules would not be updated. But on git
clone, these folders remain empty. So, if one checks the existence
of some file that should obviously be present, all would be fine.
Hence, the check is modified.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>